CAT-164 refactoring. Created MonitoringRequest class and migrated to json-b
